pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["devvourer <b.toktobekov@legres.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
